+++ tests/test_execute.py
@@ -420,10 +420,10 @@ def test_custom_node_as_product(runner, tmp_path, product_def, return_def):
     import pickle
     from typing import Any
     from typing import Annotated
-    import attrs
+    from dataclasses import dataclass
     from pytask import Product
 
-    @attrs.define
+    @dataclass
     class PickleNode:
         path: Path
         name: str = ""
@@ -750,10 +750,10 @@ def test_errors_during_loading_nodes_have_info(runner, tmp_path):
     from __future__ import annotations
     from pathlib import Path
     from typing import Any
-    import attrs
+    from dataclasses import dataclass
     import pickle
 
-    @attrs.define
+    @dataclass
     class PickleNode:
         name: str
         path: Path